fix: InviteModal 호출하는 부분 groupListPage로 이동
