+++ src/components/modals/inviteModal.module.scss
@@ -35,6 +35,7 @@
           width: 100%;
           height: 100%;
           border-radius: 50%;
+          background-color: var(--white);
         }
       }
     }
